// equa-demo-master/src/modules/app/components/routes/types.ts
export enum Paths {
  root = '/',
  login = '/login',
  register = '/register',
  terms = '/terms',
  confirmEmail = '/confirm-email/:token',
  forgotPassword = '/forgot-password',
  resetPassword = '/reset-password/:code',
  resetEmail = '/reset-email',
  changePassword = '/change-password',
  userProfile = '/user-profile',
  // UserSettings = '/user-settings',
  EditUser = '/edit-user',
  orgCreate = '/organization-create',
  orgImport = '/Organization-import',
  purchase = '/purchase',
  CompletePurchase = '/cart-summary',
  Portfolio = '/portfolio',
  Actions = '/actions',
  ActionsResults = '/action-results',
  ActionRequest = '/action-request',
  ActionDetails = '/action-details',
  Dashboard = '/dashboard',
  TeamMember = '/team-member',
  AddTeamMembers = '/add-team-member',
  teamMembersProfile = '/team-members-profile',
  editTeamMembersProfile = '/edit-team-members-profile',
  capTable = '/cap-table',
  sharesTransfer = '/shares-transfer',
  draftUnits = '/draft-units',
  Documents = '/documents',
  OperatingAgreement = '/operating-agreement',
  EditOperatingAgreement = '/edit-operating-agreement',
  CreateOperatingAgreement = '/create-operating-agreement',
  NewSubscription='/cart',
  // myReferrals = '/my-referrals',
  auth = '/auth',
  userLogin = '/auth/login',
  verify = '/auth/verify',
  userRegister = '/auth/register',
  userForgot = '/auth/forgot',
  userMeeting = '/auth/meeting',
  userThanks = '/auth/thank-you',
  startUp = '/startup',
  dataRooms = '/data-rooms',
  dataRoom = '/data-room',
  createDataRoom = '/create-data-room'
}